requirements.txt
av==9.0.0
einops==0.4.1
ftfy==6.1.1
fvcore==0.1.5.post20220305
h5py==3.6.0
hjson==3.0.2
huggingface-hub==0.6.0
hydra-core==1.1.2
iopath==0.1.9
jsonlines==3.0.0
matplotlib==3.5.1
mkl-service==2.4.0
ninja==1.10.2.3
omegaconf==2.1.2
opencv-python==4.5.5.64
pandas==1.3.5
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
Pillow==9.0.1
pyarrow==7.0.0
pycocoevalcap==1.2
pycocotools==2.0.4
pytorch-transformers==1.2.0
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
sklearn==0.0
tensorboard==2.8.0
timm==0.4.5
tqdm==4.63.0
transformers==4.19.1
cloudpickle
protobuf==3.19